refactor: moves tips to a separate component
